requirements.txt
pyserial
typer